Crate flame

Structs

Note
A note for use in debugging.

Span
A named timespan.

SpanGuard

Thread
A collection of events that happened on a single thread.

Functions

clear
Clears all of the recorded info that Flame has tracked.

commit_thread

debug
Prints all of the frames to stdout.

dump_html

dump_stdout

end
Ends the current Span and returns the number of nanoseconds that passed.

end_collapse
Ends the current Span and returns the number of nanoseconds that passed.

end_with
Ends the current Span and returns a given result.

note
Records a note on the current Span.

span_of
Starts and ends a Span that lasts for the duration of the function f.

spans
Returns a list of spans from the current thread

start
Starts a new Span

start_guard
Starts a Span and also returns a SpanGuard.

threads

Type Definitions

StrCow