Struct flame::Frame

pub struct Frame {
    pub roots: Vec<Span>,
    // some fields omitted
}

A chunk of spans that are meant to be grouped together into a "frame".

The naming (and a possible usecase) comes from gaming, where a bunch of logic and rendering happens repeatedly every "frame". When developing a game, a flamegraph can be used to analyse and debug performance issues when you see that a particular frame is oddly shaped.

If you don't have any sort of repeatable logic that you'd like to show off in your flamegraph, using a single frame is totally acceptable.

Fields

roots

A list of spans contained inside this frame.