Struct flame::Span

pub struct Span {
    pub name: StrCow,
    pub start_ns: u64,
    pub end_ns: u64,
    pub delta: u64,
    pub depth: u16,
    pub children: Vec<Span>,
    pub notes: Vec<Note>,
    // some fields omitted
}

A named timespan.

The span is the most important feature of Flame. It denotes a chunk of time that is important to you.

The Span records * Start and stop time * A list of children (also called sub-spans) * A list of notes

Fields

name

The name of the span

start_ns

The timestamp of the start of the span

end_ns

The timestamp of the end of the span

delta

The time that ellapsed between start_ns and end_ns

depth

How deep this span is in the tree

children

A list of spans that occurred inside this one

notes

A list of notes that occurred inside this span