Crate flame

Structs

Frame
A chunk of spans that are meant to be grouped together into a "frame".

Note
A note for use in debugging.

Span
A named timespan.

SpanGuard

Functions

clear
Clears all of the recorded info that Flame has tracked.

debug
Prints all of the frames to stdout.

dump_html

dump_stdout

dump_svg

end
Ends the current Span and returns the number of nanoseconds that passed.

end_collapse
Ends the current Span and returns the number of nanoseconds that passed.

frames
Returns a list of frames

next_frame
Starts a new frame.

note
Records a note on the current Span.

span_of
Starts and ends a Span that lasts for the duration of the function f.

start
Starts a new Span

start_guard
Starts a Span and also returns a SpanGuard.

Type Definitions

StrCow